Update to Latest upstream master

+++ src/main/java/cn/nukkit/nbt/stream/BufferedRandomAccessFile.java
@@ -17,7 +17,10 @@
  */
 package cn.nukkit.nbt.stream;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.RandomAccessFile;
 import java.util.Arrays;
 
 
@@ -31,7 +34,7 @@ import java.util.Arrays;
  * functioning of the <code>RandomAccessFile</code> methods that are not
  * overridden here relies on the implementation of those methods in the
  * superclass.
- * Author : Avinash Lakshman ( alakshman@facebook.com) & Prashant Malik ( pmalik@facebook.com )
+ * Author : Avinash Lakshman ( alakshman@facebook.com) &amp; Prashant Malik ( pmalik@facebook.com )
  */
 
 public class BufferedRandomAccessFile extends RandomAccessFile
@@ -383,7 +386,7 @@ public class BufferedRandomAccessFile extends RandomAccessFile
                 }
             }
         }
-        this.buff_[(int) (this.curr_ - this.lo_)] = (byte) b;
+        this.buff_[(int) (this.curr_ - this.lo_)] = b;
         this.dirty_ = true;
     }
